ref(querybuilder): Move BaseQueryBuilder out
- Moves BaseQueryBuilder to its own file
